auth test wip
