+++ apps/api-server/src/testing/index.ts
@@ -1,1 +1,1 @@
-export * from './server.js'
+export * from './environment.js'
